requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
pydantic==2.7.1
tqdm==4.66.4
umap-learn==0.5.6
hdbscan==0.8.33
